requirements.txt
crewai>=0.28.0
crewai-tools>=0.1.0
openai>=1.30.0
langchain-openai>=0.1.0
streamlit>=1.35.0
python-dotenv>=1.0.0
